requirements.txt
Flask>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.2.0
Flask-SQLAlchemy>=3.1.0
Flask-Babel>=4.0.0
WTForms>=3.1.0
email-validator>=2.0.0
Pillow>=10.0.0
Werkzeug>=3.0.0
SQLAlchemy>=2.0.0
Babel>=2.14.0
python-dotenv>=1.0.0
reportlab>=4.0.0
psutil>=5.9.0
pysmb>=1.2.0
